requirements.txt
python-docx>=0.8.11
python-pptx>=0.6.21
requests>=2.28.0
duckduckgo-search>=3.0.0
Pillow>=9.0.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
